python: rename secp256k1proto -> secp256k1lab
The commit was created with the following two commands:

$ git mv python/secp256k1proto python/secp256k1lab
$ sed -i s/secp256k1proto/secp256k1lab/g $(git ls-files)
