+++ python/.ruff.toml
@@ -1,4 +1,4 @@
 [format]
 # Exclude the file taken from Bitcoin Core for now, in case we want to
 # contribute our changes back to upstream.
-exclude = ["secp256k1proto/secp256k1.py"]
+exclude = ["secp256k1lab/secp256k1.py"]
