+++ python/chilldkg_ref/chilldkg.py
@@ -11,10 +11,10 @@ their arguments and return values, and the exceptions they raise; see also the
 from secrets import token_bytes as random_bytes
 from typing import Any, Tuple, List, NamedTuple, NewType, Optional, NoReturn, Dict
 
-from secp256k1proto.secp256k1 import Scalar, GE
-from secp256k1proto.bip340 import schnorr_sign, schnorr_verify
-from secp256k1proto.keys import pubkey_gen_plain
-from secp256k1proto.util import bytes_from_int
+from secp256k1lab.secp256k1 import Scalar, GE
+from secp256k1lab.bip340 import schnorr_sign, schnorr_verify
+from secp256k1lab.keys import pubkey_gen_plain
+from secp256k1lab.util import bytes_from_int
 
 from .vss import VSSCommitment
 from . import encpedpop
