+++ python/chilldkg_ref/encpedpop.py
@@ -1,9 +1,9 @@
 from typing import Tuple, List, NamedTuple, NoReturn
 
-from secp256k1proto.secp256k1 import Scalar, GE
-from secp256k1proto.ecdh import ecdh_libsecp256k1
-from secp256k1proto.keys import pubkey_gen_plain
-from secp256k1proto.util import int_from_bytes
+from secp256k1lab.secp256k1 import Scalar, GE
+from secp256k1lab.ecdh import ecdh_libsecp256k1
+from secp256k1lab.keys import pubkey_gen_plain
+from secp256k1lab.util import int_from_bytes
 
 from . import simplpedpop
 from .util import (
