+++ python/chilldkg_ref/simplpedpop.py
@@ -1,8 +1,8 @@
 from secrets import token_bytes as random_bytes
 from typing import List, NamedTuple, NewType, Tuple, Optional, NoReturn
 
-from secp256k1proto.bip340 import schnorr_sign, schnorr_verify
-from secp256k1proto.secp256k1 import GE, Scalar
+from secp256k1lab.bip340 import schnorr_sign, schnorr_verify
+from secp256k1lab.secp256k1 import GE, Scalar
 from .util import (
     BIP_TAG,
     FaultyParticipantOrCoordinatorError,
