+++ python/tests.sh
@@ -19,6 +19,6 @@ ruff check --quiet || true
 ruff format --diff --quiet || true
 mypy --no-error-summary . || true
 # Be more strict in the reference code
-mypy --no-error-summary --strict --untyped-calls-exclude=secp256k1proto -p chilldkg_ref --follow-imports=silent || true
+mypy --no-error-summary --strict --untyped-calls-exclude=secp256k1lab -p chilldkg_ref --follow-imports=silent || true
 
 python3 tests.py
